Move all interfaces into interface file
