--- src/lib/SteamClient.ts
@@ -1,7 +1,0 @@
-export interface BatteryState {
-  bHasBattery: boolean,
-  bShutdownRequested: boolean,
-  eACState: number,
-  flLevel: number,
-  nSecondsRemaining: number,
-}
